update components, add new blocks
