--- shared/pages/NewWallet/Buttons/ButtonCreate/ButtonCreate.scss
@@ -1,10 +1,0 @@
-.buttenCreate {
-  position: absolute;
-  width: 179px;
-  height: 44px;
-  left: 1185px;
-  top: 40px;
-  z-index: 200;
-  background: #FFFFFF;
-  border-radius: 3px;
-}
