--- shared/pages/NewWallet/Buttons/ButtonEnter/ButtonEnter.scss
@@ -1,19 +1,0 @@
-.buttonCreate {
-  position: absolute;
-  width: 134px;
-  height: 20px;
-  left: 1015px;
-  top: 52px;
-  z-index: 200;
-
-  font-family: Inter UI;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-  font-size: 16px;
-  text-align: right;
-
-  color: #FFFFFF;
-
-  opacity: 0.8;
-}
