+++ shared/pages/NewWallet/Components/Buttons/ButtonCreate/ButtonCreate.scss
@@ -1,0 +1,7 @@
+.buttenCreate {
+  z-index: 200;
+  background: #FFFFFF;
+  border-radius: 3px;
+  width: 179px;
+  height: 44px;
+}
