+++ shared/pages/NewWallet/Components/Buttons/ButtonEnter/ButtonEnter.js
