+++ shared/pages/NewWallet/Components/Buttons/ButtonEnter/ButtonEnter.scss
@@ -1,0 +1,9 @@
+.buttonCreate {
+  z-index: 200;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  font-size: 16px;
+  color: rgba(255, 255, 255, 0.8);
+  margin-right: 32px;
+}
