+++ shared/pages/NewWallet/Components/CurrencySlider/CurrencySlider.scss
@@ -1,18 +1,54 @@
 .availableCurrencies {
-  position: absolute;
-  width: 380px;
-  height: 29px;
-  left: 160px;
-  top: 490px;
-
-  font-family: Inter UI;
   font-style: normal;
   font-weight: 600;
-  line-height: 29px;
   font-size: 20px;
-
   color: #000000;
-  padding-top: 45px;
-  border-top: solid 0.5px #A9A9A9;
+}
 
+.currencyList {
+  list-style: none;
+  padding: 0;
+  display: flex;
+  align-items: center;
+  margin-top: 40px;
+}
+.currencyListItem {
+  margin-right: 40px;
+  display: inline-block;
+  position: relative;
+}
+
+.currencyListWrap {
+  position: relative;
+}
+
+.fireItem {
+  position: absolute;
+  width: 30px;
+  justify-content: center;
+  align-items: center;
+  height: 30px;
+  display: flex;
+  border-radius: 50%;
+  right: -20px;
+  top: -20px;
+  z-index: 111;
+  background: #fff;
+}
+
+.currencyAdd {
+  position: absolute;
+  width: 56px;
+  height: 56px;
+  background: #FFFFFF;
+  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.1);
+  border-radius: 50%;
+  color: #000000;
+  display: flex;
+  font-weight: 100;
+  justify-content: center;
+  font-size: 35px;
+  position: absolute;
+  right: 30px;
+  top: -7px;
 }
