+++ shared/pages/NewWallet/Components/NewHeader/NewHeader.js
@@ -3,9 +3,6 @@ import React, { Component } from 'react'
 import styles from './NewHeader.scss'
 import cssModules from 'react-css-modules'
 import Logo from 'components/Logo/Logo'
-import ButtonEnter from '../Buttons/ButtonEnter/ButtonEnter'
-import ButtonCreate from '../Buttons/ButtonCreate/ButtonCreate'
-
 import Menu from '../Menu/Menu'
 
 
@@ -18,12 +15,6 @@ export default class NewHeader extends Component {
           <Logo />
         </div>
         <Menu />
-        <ButtonEnter onClick={() => console.log('asd')}>
-          Enter to wallet
-        </ButtonEnter>
-        <ButtonCreate onClick={() => console.log('asd')}>
-          Create new wallet
-        </ButtonCreate>
       </div>
     )
   }
