+++ shared/pages/NewWallet/Components/NewHeader/NewHeader.scss
