+++ shared/pages/NewWallet/Components/Promo/Promo.scss
@@ -1,0 +1,10 @@
+.promo {
+  width: 67.3%;
+  background: #F5F5F5;;
+}
+.promoWrap {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  left: 160px;
+}
