+++ shared/pages/NewWallet/Components/PromoText/PromoText.scss
@@ -1,13 +1,14 @@
 .promoText {
-  position: absolute;
   width: 525px;
-  height: 67px;
-  left: 160px;
-  top: 200px;
-
-  font-family: Inter UI;
   font-style: normal;
   font-weight: bold;
   line-height: normal;
   font-size: 56px;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+  padding-bottom: 40px;
+  margin-bottom: 50px;
+}
+
+.promoTextColored {
+  color: #6144E5;
 }
