+++ shared/pages/NewWallet/Components/Side/Side.scss
@@ -1,0 +1,9 @@
+.side {
+  width: 33.3%;
+  background: #6144E5;
+}
+.buttonsWrap {
+  display: flex;
+  margin-top: 40px;
+  justify-content: center;
+}
