--- shared/pages/NewWallet/Components/SwapWindow/ExchangeBlock.scss
@@ -1,11 +1,0 @@
-.block {
-  position: absolute;
-  width: 460px;
-  height: 453px;
-  left: 820px;
-  top: 172px;
-
-  background: #FFFFFF;
-  box-shadow: 0px 0px 8px rgba(0, 0, 0, 0.1);
-  border-radius: 8px;
-}
