--- shared/pages/NewWallet/Components/SwapWindow/SwapWindow/SwapWindow.scss
@@ -1,223 +1,0 @@
-@media (min-width: 831px) {
-  .TitleHolder {
-    position: absolute;
-    width: 50%;
-    top: 87px;
-    h1, span {
-      font-size: 45px;
-    }
-    * {
-      margin: 0;
-      padding: 0;
-    }
-  }
-
-  .section {
-    padding-top: 70px;
-  }
-}
-
-.button {
-  margin-top: 20px;
-  width: 45%;
-  font-size: 14px;
-}
-
-.rowBtn {
-  display: flex;
-  justify-content: space-around;
-}
-
-.walletToggle {
-  margin-top: 10px;
-  &_site {
-    margin-bottom: 15px;
-  }
-}
-
-.anotherRecepient {
-  position: relative;
-  z-index: -1;
-  max-height: 0;
-  opacity: 0;
-  transition: max-height 0.3s ease;
-  &_active {
-    z-index: 0;
-    opacity: 1;
-    max-height: 100vh;
-    transition: opacity 0.3s 0.2s,
-                max-height 0.7s ease;
-  }
-}
-
-.walletInput {
-  margin-top: 10px;
-}
-
-.section, .widgetSection {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
-.blockVideo {
-  flex-basis: 58%;
-  iframe {
-    width: 100%;
-    height: 355px;
-  }
-}
-
-.block {
-  position: relative;
-  flex-basis: 37%;
-}
-
-.error {
-  color: red;
-  a {
-    position: relative;
-    cursor: pointer;
-    &::after {
-      display: block;
-      position: absolute;
-      content: '';
-      bottom: -1px;
-      width: 100%;
-      height: 1px;
-      background: red;
-    }
-  }
-}
-
-.row {
-  display: flex;
-  align-items: center;
-}
-
-.flipButton {
-  position: static;
-  padding: 0;
-  margin: 0px 12px -20px auto;
-  transform: rotate(90deg);
-}
-
-.withLoader {
-  display: block;
-  position: relative;
-
-}
-
-.inform {
-  color: #626262;
-  margin-top: 40px;
-}
-
-.loaderHolder {
-  display: inline-block;
-  position: relative;
-  top: 2px;
-}
-
-.additionalLoaderHolder {
-  display: block;
-  position: relative;
-  > * {
-    display: block;
-    position: static !important;
-    margin: 0 !important;
-    padding: 0 !important;
-  }
-}
-
-.extendedControls {
-  position: relative;
-  z-index: 1;
-  width: 100%;
-  background: white;
-  overflow: visible;
-  margin: 1px 0 -19px;
-  > div {
-    margin: 0;
-  }
-}
-
-.maxAmount {
-  display: inline-block;
-  position: absolute;
-  top: 0px;
-  right: 0;
-  margin-bottom: 0;
-}
-
-@include media-mobile {
-  .inform {
-    margin-top: 10px;
-  }
-  .maxAmount {
-    top: 6px;
-  }
-}
-
-@include media-tablet {
-  .blockVideo {
-    flex-basis: 50%;
-  }
-  .block {
-    flex-basis: 45%;
-  }
-  .button {
-    font-size: 12px;
-  }
-  .maxAmount {
-    font-size: 14px;
-  }
-  .TitleHolder {
-    width: 45%;
-    h1, span {
-      font-size: 35px;
-    }
-  }
-}
-@media all and (min-width: 320px) and (max-width: 830px) {
-  .section, .widgetSection {
-    margin-top: 20px;
-    justify-content: center;
-    flex-direction: column-reverse;
-    max-width: 400px;
-    margin: auto;
-  }
-  .button {
-    font-size: 12px;
-  }
-  .block {
-    margin-bottom: 80px;
-    margin-top: 60px;
-  }
-  .widgetSection {
-    padding: 0 25px;
-    overflow: hidden;
-    .flipButton {
-      left: auto;
-      right: 40px;
-      transform: translateY(12px) rotate(90deg);
-    }
-    .selectedItem {
-      width: auto;
-    }
-  }
-  .TitleHolder {
-    width: 100%;
-  }
-}
-
-@media all and (min-width: 320px) and (max-width: 400px) {
-  .button {
-    font-size: 9px;
-  }
-}
-
-@include media-widget {
-  .block {
-    margin: 0 auto;
-  }
-}
