+++ shared/pages/NewWallet/NewWallet.scss
@@ -1,7 +1,4 @@
 .container {
   display: flex;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  background: #E5E5E5;
+  height: 100vh;
 }
