--- shared/pages/NewWallet/Side/Side.js
@@ -1,19 +1,0 @@
-import React, { Component } from 'react'
-
-
-import styles from './Side.scss'
-import cssModules from 'react-css-modules'
-
-import NewHeader from '../NewHeader/NewHeader'
-
-
-@cssModules(styles)
-export default class NewContent extends Component {
-  render() {
-    return (
-      <div styleName="side">
-  
-      </div>
-    )
-  }
-}
