--- shared/pages/NewWallet/Side/Side.scss
@@ -1,8 +1,0 @@
-.side {
-  position: absolute;
-  width: 480px;
-  height: 800px;
-  left: 960px;
-  top: 0px;
-  background: #6144E5;
-}
